Allow script to find data even when imported
Several other projects either run synonymize.py or import it as a
module. But the Plant List data was referred to by a relative path that
would not work if the working directory was different from the one in
which synonymize.py was found. This fixes that by finding the path of
the file itself at run time.